Fix the last few warnings
Enables TreatWarningsAsErrors for all projects
