+++ Emby.Server.Implementations/Images/MusicGenreImageProvider.cs
@@ -8,8 +8,7 @@ using MediaBrowser.Common.Configuration;
 using MediaBrowser.Controller.Drawing;
 using MediaBrowser.Controller.Dto;
 using MediaBrowser.Controller.Entities;
-using MediaBrowser.Controller.Entities.Movies;
-using MediaBrowser.Controller.Entities.TV;
+using MediaBrowser.Controller.Entities.Audio;
 using MediaBrowser.Controller.Library;
 using MediaBrowser.Controller.Providers;
 using MediaBrowser.Model.Entities;
@@ -19,31 +18,36 @@ using MediaBrowser.Model.Querying;
 namespace Emby.Server.Implementations.Images
 {
     /// <summary>
-    /// Class GenreImageProvider.
+    /// Class MusicGenreImageProvider.
     /// </summary>
-    public class GenreImageProvider : BaseDynamicImageProvider<Genre>
+    public class MusicGenreImageProvider : BaseDynamicImageProvider<MusicGenre>
     {
         /// <summary>
         /// The library manager.
         /// </summary>
         private readonly ILibraryManager _libraryManager;
 
-        public GenreImageProvider(IFileSystem fileSystem, IProviderManager providerManager, IApplicationPaths applicationPaths, IImageProcessor imageProcessor, ILibraryManager libraryManager) : base(fileSystem, providerManager, applicationPaths, imageProcessor)
+        public MusicGenreImageProvider(IFileSystem fileSystem, IProviderManager providerManager, IApplicationPaths applicationPaths, IImageProcessor imageProcessor, ILibraryManager libraryManager) : base(fileSystem, providerManager, applicationPaths, imageProcessor)
         {
             _libraryManager = libraryManager;
         }
 
         /// <summary>
-        /// Get children objects used to create an genre image.
+        /// Get children objects used to create an music genre image.
         /// </summary>
-        /// <param name="item">The genre used to create the image.</param>
+        /// <param name="item">The music genre used to create the image.</param>
         /// <returns>Any relevant children objects.</returns>
         protected override IReadOnlyList<BaseItem> GetItemsWithImages(BaseItem item)
         {
             return _libraryManager.GetItemList(new InternalItemsQuery
             {
                 Genres = new[] { item.Name },
-                IncludeItemTypes = new[] { nameof(Series), nameof(Movie) },
+                IncludeItemTypes = new[]
+                {
+                    nameof(MusicAlbum),
+                    nameof(MusicVideo),
+                    nameof(Audio)
+                },
                 OrderBy = new[] { (ItemSortBy.Random, SortOrder.Ascending) },
                 Limit = 4,
                 Recursive = true,
