+++ Emby.Server.Implementations/Library/Resolvers/FolderResolver.cs
@@ -9,7 +9,7 @@ namespace Emby.Server.Implementations.Library.Resolvers
     /// <summary>
     /// Class FolderResolver.
     /// </summary>
-    public class FolderResolver : FolderResolver<Folder>
+    public class FolderResolver : GenericFolderResolver<Folder>
     {
         /// <summary>
         /// Gets the priority.
@@ -30,26 +30,6 @@ namespace Emby.Server.Implementations.Library.Resolvers
             }
 
             return null;
-        }
-    }
-
-    /// <summary>
-    /// Class FolderResolver.
-    /// </summary>
-    /// <typeparam name="TItemType">The type of the T item type.</typeparam>
-    public abstract class FolderResolver<TItemType> : ItemResolver<TItemType>
-        where TItemType : Folder, new()
-    {
-        /// <summary>
-        /// Sets the initial item values.
-        /// </summary>
-        /// <param name="item">The item.</param>
-        /// <param name="args">The args.</param>
-        protected override void SetInitialItemValues(TItemType item, ItemResolveArgs args)
-        {
-            base.SetInitialItemValues(item, args);
-
-            item.IsRoot = args.Parent == null;
         }
     }
 }
