+++ Emby.Server.Implementations/Library/Resolvers/PlaylistResolver.cs
@@ -16,7 +16,7 @@ namespace Emby.Server.Implementations.Library.Resolvers
     /// <summary>
     /// <see cref="IItemResolver"/> for <see cref="Playlist"/> library items.
     /// </summary>
-    public class PlaylistResolver : FolderResolver<Playlist>
+    public class PlaylistResolver : GenericFolderResolver<Playlist>
     {
         private string[] _musicPlaylistCollectionTypes =
         {
