+++ Emby.Server.Implementations/LiveTv/TunerHosts/HdHomerun/HdHomerunManager.cs
@@ -5,12 +5,10 @@
 using System;
 using System.Buffers;
 using System.Buffers.Binary;
-using System.Collections.Generic;
 using System.Globalization;
 using System.Net;
 using System.Net.Sockets;
 using System.Text;
-using System.Text.RegularExpressions;
 using System.Threading;
 using System.Threading.Tasks;
 using MediaBrowser.Common;
@@ -18,70 +16,6 @@ using MediaBrowser.Controller.LiveTv;
 
 namespace Emby.Server.Implementations.LiveTv.TunerHosts.HdHomerun
 {
-    public interface IHdHomerunChannelCommands
-    {
-        IEnumerable<(string, string)> GetCommands();
-    }
-
-    public class LegacyHdHomerunChannelCommands : IHdHomerunChannelCommands
-    {
-        private string _channel;
-        private string _program;
-
-        public LegacyHdHomerunChannelCommands(string url)
-        {
-            // parse url for channel and program
-            var regExp = new Regex(@"\/ch([0-9]+)-?([0-9]*)");
-            var match = regExp.Match(url);
-            if (match.Success)
-            {
-                _channel = match.Groups[1].Value;
-                _program = match.Groups[2].Value;
-            }
-        }
-
-        public IEnumerable<(string, string)> GetCommands()
-        {
-            if (!string.IsNullOrEmpty(_channel))
-            {
-                yield return ("channel", _channel);
-            }
-
-            if (!string.IsNullOrEmpty(_program))
-            {
-                yield return ("program", _program);
-            }
-        }
-    }
-
-    public class HdHomerunChannelCommands : IHdHomerunChannelCommands
-    {
-        private string _channel;
-        private string _profile;
-
-        public HdHomerunChannelCommands(string channel, string profile)
-        {
-            _channel = channel;
-            _profile = profile;
-        }
-
-        public IEnumerable<(string, string)> GetCommands()
-        {
-            if (!string.IsNullOrEmpty(_channel))
-            {
-                if (!string.IsNullOrEmpty(_profile)
-                    && !string.Equals(_profile, "native", StringComparison.OrdinalIgnoreCase))
-                {
-                    yield return ("vchannel", $"{_channel} transcode={_profile}");
-                }
-                else
-                {
-                    yield return ("vchannel", _channel);
-                }
-            }
-        }
-    }
-
     public sealed class HdHomerunManager : IDisposable
     {
         public const int HdHomeRunPort = 65001;
